Make sure that finufftc links to libfinufft.so
On certain systems, the default ld option --no-as-needed is overridden
by --as-needed, which means that only dynamic libraries which contain
symbols that unresolved in the main object files are included in the
DT_NEEDED fields. This happens, for example, on Ubuntu 20.04.

To resolve this, we create a temporary source file which contains a call
to a function in `libfinufft.so`, forcing the linker to include it.