Change python target to editable install
This way, pip will only insert a link to the source code instead of
installing it in a system- or user-specific location. Since the library
is linked to a hardcoded path for libfinufft.so, this makes things more
consistent in that the dynamic library and the code that uses it remains
together.