maint: reorganize directories
This should give the repository a more coherent structure, with main
(C++) code living in `src`, while all wrappers (Fortran, Python,
Matlab), etc. living in their own directories (along with examples,
tests, and so on). When relevant, these subdirectories are split
according to CPU or GPU version (as for the Python wrappers) or we have
a subdirectory for the cuda code.