cuda: fix cufinufft setup.py
After the incorporation of the cufinufft codebase into the finufft
repository, the directory structure has changed, which needs to be
reflected in the setup.py file for cufinufft. In a similar way to the
finufft setup.py, we can now specify a CUFINUFFT_DIR root directory
where setup.py will look for the `build` subdirectory. Alternatively,
it'll assume that setup.py is being called from its standard location in
the repository and find the build directory using a relative path.