cuda: specify py versions for wheels
Since manylinux2014 includes the prerelease version of py312 (which we
are not yet compatible with), we need to exclude this from the set of
Python versions for which we are compiling wheels. While we're at it,
this commit specifies a list of versions instead that we can update as
we go along.