cuda: move all internal api calls off the default stream
Provides the basic groundwork to allow for experimenting with different strategies for
improving concurrency via async operations.

The d_plan now maintains a list of stream pointers (up to d_plan->nstreams), though it's
hard-coded to only have one entry currently. It also contains a "pointer" called "curr_stream"
to specify the current stream to do an operation on, which is currently fixed at 0.