bugfix: subsequent calls to makeplan now set number of fftw threads
Before, multiple calls to our makeplan routines would always use the same number of FFTW
threads as the first makeplan call. If the user had set the number of threads themselves, it
would override that as well. This commit makes each call unique as well as returning the value
to whatever was set before the makeplan call.