Craete dummy source when building cufinufft C ext
Similarly to what we do for finufft. The issue is that the default gcc
config on certain OSes (Ubuntu 20.04, Rocky 8), libraries are not linked
unless their symbols are explicitly referenced in the source. Since we
have no source, no libraries are linked. The solution is to create a
dummy source that just calls a single function
`cufinufft_default_opts`.