ci: run macOS wheel building from repo root
Running some of these commands (notably `pytest`) from `python/finufft`
confuses the interpreter as it tries to look for `finufft.finufftc` in
the source directory (`python/finufft/finufft`) instead of
`site-packages`. This behavior seems to be related to replacing `imp`
with `importlib`. At any rate, the solution is to stay in the repo root
and run all the commands from there, prefixing with `python/finufft`
when necessary.