[BugFix] Robustize cuFINUFFT Python guru interface.
As per the docs:
* finufft.Plan(n_modes_or_dim) accepts (int, tuple[int])
* cufinufft.Plan(n_modes) accepts (int, tuple[int])

In practice `n_modes_or_dim` is transformed internally to a NumPy array, hence providing any iterable to finufft.Plan() works.
However `n_modes` in cufinufft.Plan() is more restrictive: it works with tuple[int], fails planning if list[int], and wrongfully passes planning if ndarray[int], but then seg-faults at execute().

This commit increases the robustness of cufinufft.Plan() to match the flexibility of the CPU interface.