py: skip fallback tests in CI
For some reason, these tests work locally, but in CI, the library is
imported as usual, even with the patched function. Perhaps this has
something to do with the fact that the module is already loaded, so
`import finufft` is a no-op (this is evidenced by the fact that any
`print` statement in `finufft._finufft` is not detected in the output).