Fix cufinufft fallback (#536)
* py: fix bug in cufinufft library fallback

* py: add tests for fallback library loading

* py: skip fallback tests in CI

These tests checks that the fallback mechanism is working properly. In
other words, if there is no bundled library, it should look in the
system path (e.g. `LD_LIBRARY_PATH` on Unix systems) and try to load it
from there.

For some reason, these tests work locally, but in CI, the library is
imported as usual, even with the patched function. Perhaps this has
something to do with the fact that the module is already loaded, so
`import finufft` is a no-op (this is evidenced by the fact that any
`print` statement in `finufft._finufft` is not detected in the output).