ci: add workflow for building CUDA py wheels
This sets up a workflow that will build the cuFINUFFT (i.e., GPU) Python
wheels using `cibuildwheel` (which in turn relies on a Docker image that
runs `build` and `auditwheel` on a `manylinux2014` image). Care is taken
so that the necessary CUDA build tools are installed and found by
`cmake`. The resulting wheels are then uploaded as artifacts to GitHub.