Fix julia reference clashes
We currently have three conflicting Julia references in the rST sources.
The first is the section heading in `julia.rst` (created using
`.. _julia:`). The second is the `julia` link in `users.rst`, while the
third is the `Julia` link in `julia.rst`. It turns out that if you write
your link on the format `text <https://url/to/page>`_, Sphinx will
generate a target with the name `text` that you then can use for repeat
links. To generate an *anonymous link* (that doesn't create such a
target), you have to end with two underscores, like so: `text
<https://url/to/page>`__. To fix this, I made these links anonymous.

should consider doing the same for other link to avoid this in the future.