Minor fixes to docstrings and docs (#778)
* Fix Python docstrings

Some formatting issues in the auto-generated docstrings made the
parameters not appear properly when calling `help` and in the
rendered HTML. These are now correct.

* Remove unnused static folder option

* Fix link to NFFT.jl

* Fix links in related.rst

Should be to documents, not sections in documents.

* Fix julia reference clashes

We currently have three conflicting Julia references in the rST sources.
The first is the section heading in `julia.rst` (created using
`.. _julia:`). The second is the `julia` link in `users.rst`, while the
third is the `Julia` link in `julia.rst`. It turns out that if you write
your link on the format `text <https://url/to/page>`_, Sphinx will
generate a target with the name `text` that you then can use for repeat
links. To generate an *anonymous link* (that doesn't create such a
target), you have to end with two underscores, like so: `text
<https://url/to/page>`__. To fix this, I made these links anonymous.

should consider doing the same for other link to avoid this in the future.

* Fix other link target clashes

* Fix bulleted lits in performance.rst

Should not be indented and needs a blank line in between.

* Mark out as optional for finufft*d3

* Add docstrings to cufinufft*d3